chore: migrate text and error to TS
