+++ packages/cli-testing-library/src/index.js
@@ -27,11 +27,11 @@ if (
 export * from './config'
 export * from './helpers'
 export * from './events'
-export * from './get-queries-for-instance'
+export * from './get-queries-for-instance.js'
 export * from './matches'
 export * from './pure'
 export * from './query-helpers'
-export * from './queries'
+export * from './queries/index.js'
 export * from './mutation-observer'
 export * from './wait-for'
 export * from './user-event'
