+++ packages/cli-testing-library/src/queries/error.ts
@@ -1,7 +1,27 @@
-import {GetErrorFunction, QueryByText} from '../../types'
+import {GetErrorFunction, Matcher, SelectorMatcherOptions, waitForOptions} from '../../types'
 import {fuzzyMatches, matches, makeNormalizer, buildQueries} from './all-utils'
+import {TestInstance} from "../types";
 
-const queryByErrorBase: QueryByText = (
+export type QueryByError<T extends TestInstance = TestInstance> = (
+  instance: TestInstance,
+  id: Matcher,
+  options?: SelectorMatcherOptions,
+) => T | null
+
+export type GetByError<T extends TestInstance = TestInstance> = (
+  instance: TestInstance,
+  id: Matcher,
+  options?: SelectorMatcherOptions,
+) => T
+
+export type FindByError<T extends TestInstance = TestInstance> = (
+  instance: TestInstance,
+  id: Matcher,
+  options?: SelectorMatcherOptions,
+  waitForElementOptions?: waitForOptions,
+) => Promise<T>
+
+const queryByErrorBase: QueryByError = (
   instance,
   text,
   {exact = false, collapseWhitespace, trim, normalizer, stripAnsi} = {},
@@ -18,12 +38,28 @@ const queryByErrorBase: QueryByText = (
   else return null
 }
 
-const getMissingError: GetErrorFunction<[unknown]> = (c, text) =>
+const getMissingError: GetErrorFunction<[unknown]> = (_c, text) =>
   `Unable to find an stdout line with the text: ${text}. This could be because the text is broken up by multiple lines. In this case, you can provide a function for your text matcher to make your matcher more flexible.`
 
-const [queryByErrorWithSuggestions, getByError, findByError] = buildQueries(
+const [_queryByErrorWithSuggestions, _getByError, _findByError] = buildQueries(
   queryByErrorBase,
   getMissingError,
 )
 
-export {queryByErrorWithSuggestions as queryByError, getByError, findByError}
+export function getByError<T extends TestInstance = TestInstance>(
+  ...args: Parameters<GetByError<T>>
+): ReturnType<GetByError<T>> {
+  return _getByError<T>(...args);
+}
+
+export function queryByError<T extends TestInstance = TestInstance>(
+  ...args: Parameters<QueryByError<T>>
+): ReturnType<QueryByError<T>> {
+  return _queryByErrorWithSuggestions<T>(...args);
+}
+
+export function findByError<T extends TestInstance = TestInstance>(
+  ...args: Parameters<FindByError<T>>
+): ReturnType<FindByError<T>> {
+  return _findByError<T>(...args);
+}
